Moving some functions to llm utils.py
